@@ -1,48 +1,51 @@
 {
-  "name": "angularv9",
+  "name": "v9",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
     "start": "ng serve",
     "build": "ng build",
-    "lint": "ng lint",
     "test": "npm run test:dev && npm run test:prod",
-    "test:dev": "ng build && cypress-runner run --path dist/v8",
-    "test:prod": "ng build --prod && cypress-runner run --path dist/v8"
+    "test:dev": "ng build && cypress-runner run --path dist/v9",
+    "test:prod": "ng build --prod && cypress-runner run --path dist/v9",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~9.0.0-rc.1",
-    "@angular/compiler": "~9.0.0-rc.1",
-    "@angular/core": "~9.0.0-rc.1",
-    "@angular/platform-browser": "~9.0.0-rc.1",
-    "@angular/platform-browser-dynamic": "~9.0.0-rc.1",
-    "@uirouter/angular": "latest",
+    "@angular/animations": "~9.0.3",
+    "@angular/common": "~9.0.3",
+    "@angular/compiler": "~9.0.3",
+    "@angular/core": "~9.0.3",
+    "@angular/forms": "~9.0.3",
+    "@angular/platform-browser": "~9.0.3",
+    "@angular/platform-browser-dynamic": "~9.0.3",
+    "@angular/router": "~9.0.3",
     "@uirouter/cypress-runner": "^1.0.10",
-    "rxjs": "~6.5.3",
+    "@uirouter/angular": "latest",
+    "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.0-rc.1",
-    "@angular/cli": "~9.0.0-rc.1",
-    "@angular/compiler-cli": "~9.0.0-rc.1",
-    "@angular/language-service": "~9.0.0-rc.1",
+    "@angular-devkit/build-angular": "~0.900.4",
+    "@angular/cli": "~9.0.4",
+    "@angular/compiler-cli": "~9.0.3",
+    "@angular/language-service": "~9.0.3",
     "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.3.8",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.4.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.6.4"
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.7.5"
   }
 }
